refactor: centralize frequency profile
